Улучшен скелет для юнит тестов
