+++ tests/NoticeRawTest.php
@@ -5,6 +5,7 @@ use PHPUnit\Framework\TestCase;
 class NoticeRawTest extends TestCase {
 
     public function testNotice() {
+        // @todo
         $stack = [];
         $this->assertSame(0, count($stack));
